fix: Move `/core/` to main landing page, and remove --`core` on CLI
