+++ docs/src/components/landing-studio/Banner.js
