+++ docs/src/components/landing-studio/GithubStars.js
