+++ docs/src/components/landing-studio/HeroVideo.js
