+++ docs/src/components/landing-studio/PricingTable.js
