+++ docs/src/components/landing/GetStartedButtons.js
@@ -5,9 +5,10 @@ import Button from '@mui/material/Button';
 import KeyboardArrowRightRounded from '@mui/icons-material/KeyboardArrowRightRounded';
 import { Link } from '@mui/docs/Link';
 import NpmCopyButton from 'docs/src/components/action/NpmCopyButton';
+import GithubStars from '../landing-studio/GithubStars';
 
 export default function GetStartedButtons(props) {
-  const { installation, primaryLabel, primaryUrl, secondaryLabel, secondaryUrl, ...other } = props;
+  const { installation, primaryLabel, primaryUrl, ...other } = props;
   return (
     <React.Fragment>
       <Box
@@ -25,7 +26,6 @@ export default function GetStartedButtons(props) {
           href={primaryUrl}
           component={Link}
           noLinkStyle
-          size="large"
           variant="contained"
           endIcon={<KeyboardArrowRightRounded />}
           sx={{
@@ -36,20 +36,7 @@ export default function GetStartedButtons(props) {
         >
           {primaryLabel}
         </Button>
-        {secondaryUrl ? (
-          <Button
-            href={secondaryUrl}
-            component={Link}
-            noLinkStyle
-            target={'_blank'}
-            variant="outlined"
-            size="large"
-            color="secondary"
-            endIcon={<KeyboardArrowRightRounded />}
-          >
-            {secondaryLabel}
-          </Button>
-        ) : null}
+        <GithubStars />
       </Box>
       <NpmCopyButton installation={installation} sx={{ mt: 2 }} />
     </React.Fragment>
